[tripping-while-skipping] [Bug Fix] Ensure that we can handle missing rows in the user inputs for the Matrix Widget
Issue: Entering information in the first row of a Matrix Widget, and then skipping an entire row to interact with later rows would result in the page erroring out.

The way we're saving the user input for the Matrix Widget was causing issues with our validation logic, as it was attempting to access a property that did not exist. This minor PR ensures that our validation logic properly accounts for the user input format, and also fixes a test to properly mock how our user input is actually saved.

Issue: LEMS-3266

Test Plan:
- Tests pass
- Improved test