[Radio] Refactor multiple choice option (logic and UI updates) (#2697)
## Summary:
This PR refactors the logic, HTML, and CSS for the choice option. This component was simplified to just handle the display of the option content and its styling. Accessibility features will be added after the component is refactored so that it can be applied in a holistic manner across the various files.

Issue: LEMS-3168

## Test plan:
1. Launch Storybook locally
1. Navigate to Perseus => Widgets => Radio => Widget Internal Components => [Single Select](http://localhost:6006/?path=/story/perseus-widgets-radio-widget-internal-components-choice--single-select)
1. Visually, the different states match the new specs.
   - Note that clicking anywhere in the option will toggle the indicator (for the options that are not in review mode).

Author: mark-fitzgerald

Reviewers: ivyolamit, SonicScrewdriver, mark-fitzgerald, anakaren-rojas, nishasy

Required Reviewers:

Approved By: SonicScrewdriver, ivyolamit

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2697