[remove-expression-error-tooltip] Remove the expression widget validation tooltip
this is an anti-pattern, no other widget (to my knowledge) has this type of realtime feedback - feedback is generally provided after submit - including for this invalid use case. It is causing a styling issue with assessment and a pattern we want to move away from, so removing this code.

Issue: LEMS-3292

Test plan:
1. go to an expression widget
2. enter an invalid input such as just a single hyphen
3. observe that there is no immediate feedback that your answer is invalid
(upon submitting, user should be warned their answer is invalid and not scored incorrectly, maintaining the existing behavior)