[ig-point-0] [Interactive Graph][Point] Allow points on Point graphs to sit on edge of graph
We received a request from content authors to allow users to move a point to
the edge of the graph, specifically for the case that a graph start at 0 in
both the x and y direction (like in physics) and a point needs to sit at (0, 0).

To allow this, I'm making it so that the bounds are not inset and the overflow
is not hidden ONLY FOR POINT GRAPHS.

I'll explore other graph types if specifically requested, but this is not possible
for most graph types because of their overflow.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3213

Test plan:
`pnpm jest packages/perseus/src/widgets/interactive-graphs/reducer/interactive-graph-reducer.test.ts`

Storybook
- http://localhost:6006/?path=/story/perseuseditor-widgets-interactive-graph--interactive-graph-point
- Move a point to the corner of the graph
- Confirm that it can sit at the corner of the graph, fully visible
- Confirm it can't go past the bounds of the graph