[ig-arrows-widget] [BLOCKED - do not review] [Interactive Graph] Add `boundedSides` field
Working on adding a way for content authors to specify which axis arrows to remove
for graphs that need to be bounded in some direction. This can be any combination
of x min, x max, y min, and y max directions.

In this PR, I'm adding a `boundedSides` field to the Interactive Graph schema, and
I'm connecting it to the widget. The editor updates will be in a following PR.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3214

Test plan:
`pnpm jest packages/perseus/src/widgets/interactive-graphs/interactive-graph-question-builder.test.ts`
Check this PR for the new regression tests, check future PRs for regression failures

Storybook
- http://localhost:6006/?path=/story/perseus-widgets-interactive-graph-visual-regression-tests--bounded-x-min
- http://localhost:6006/?path=/story/perseus-widgets-interactive-graph-visual-regression-tests--bounded-x-max
- http://localhost:6006/?path=/story/perseus-widgets-interactive-graph-visual-regression-tests--bounded-y-min
- http://localhost:6006/?path=/story/perseus-widgets-interactive-graph-visual-regression-tests--bounded-y-max
- http://localhost:6006/?path=/story/perseus-widgets-interactive-graph-visual-regression-tests--bounded-all-sides