docs(changeset): Bugfix to ensure Interactive Graph tooltips display points with the same granularity as the snapStep (#2728)
## Summary:
We've had bugs in Interactive Graph where certain labels are hitting the Javascript floating point issues. 

LEMS-3241
This PR fixes the floating point bug related to the point labels by ensuring that we're always showing the same number of significant digits as the snap step. This ensures that the labels are always a consistent length.

LEMS-3216
This PR also fixes LEMS-3216 with a very similar solution, in order to ensure the tick steps all have the same number of sig figs. 

Additionally, there's a minor bug-fix that was snuck in here to help fix a regression in a testing tool that's necessary to properly recreate some of these issues. All that was required to fix the issue was wrapping the new UserInputManager around the renderer in the interactive graph regression story file. 

Issue: LEMS-3241, LEMS-3216

## Test plan:
- New graph view for manual testing

Author: SonicScrewdriver

Reviewers: SonicScrewdriver, ivyolamit, nishasy

Required Reviewers:

Approved By: ivyolamit, nishasy

Checks: ✅ 9 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2728