Detect item splitting function changes in Github Action (#2719)
## Summary:

This PR builds on the work in #2711 by adding another Github Action job to check for changes to our item splitting (for answerless data) logic. It does this by:

  * adding a new export to `@khanacademy/perseus-core` named `item-splitting` (so you'd import it as `import {splitPerseusItem} from "@khanacademy/perseus-core/item-splitting";`).  This export contains _all_ of the item splitting code in a single file so we can be sure if it changes, we know the logic has changed.
  * adding a Github Action job in `node-ci.yml` to build the current PR's code and its base branch code and compare the two item-splitting bundles.
  * refactors the schema change logic into a new, composite Github action and uses it for both schema change checks and item splitting logic checks

Issue: LEMS-3305

## Test plan:

This PR cannot fully test the split function change detection because it introduces an `"exports"` to the `perseus-core` package. This export causes _all_ of the split logic to be bundled into the new export and we use that as the basis for checking for changes. 

So, my plan is to review this PR and land it.  Then start a new PR and test/verify the change detection. I can also tweak the action and diff rendering in that new PR because by that time the base branch will be building the new `"exports"` bundle. 

Finally, the two "change detection" jobs are failing in the "post" step (see screenshot). I am fairly sure that it's because it's trying to use the composite action from the base branch to cleanup but it doesn't exist there so it fails. My hunch is that this error will also resolve once I land this PR and begin testing in a new one. 

<img width="1153" height="112" alt="image" src="https://github.com/user-attachments/assets/68428be9-f161-4e93-8b8c-b6698c029484" />

Author: jeremywiebe

Reviewers: benchristel, jeremywiebe, mark-fitzgerald, ivyolamit, Myranae

Required Reviewers:

Approved By: mark-fitzgerald

Checks: ❌ 2 checks failed, ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2719