[remove_free_response_forced_bold] Remove forced bold from Free Response question
This commit changes the styling on the Free Respone Widget to no longer
force bold styling of the question text. Instead, we will allow the
content creators to use Markdown to bold all or part of the question, if
they want to.

Issue: https://khanacademy.atlassian.net/browse/AX-1596

Test plan:
- Open the Free Response Widget in Storybook.
- View the question on the Primary story.
- Verify that it is not rendered bold.
- In the browser DevTools, inspect the question text.
- Verify that it is not surrounded by `<strong>` tags.
- In the browser DevTools, inspect the question text of the "Bolded
  Question" story.
- Verify that it is surrounded by `<strong>` tags.