[image-mobile] [Image] (UX) Remove the custom mobile styling for title and caption
As part of an archaic mobile design rehaul, the image widget had the
title added to its caption section and bolded. The code for this is
somewhat convoluted, and this separate design is not necessary.

Uniting the desktop/mobile code by removing the custom mobile section.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3381

Test plan:
Storybook:
- http://localhost:6006/?path=/docs/widgets-image--docs
- Click on the mobile preview switch, and confirm it's the same as desktop