[Radio] Adjust horizontal scroll fade bar color to use background color of container (#2816)
## Summary
The Graded Group Set has a background color that is not the page color. When a radio widget is included in this widget, the fade bars used for horizontal scrolling show as white against the gray background. The radio widget doesn't know that it is being displayed in the graded group set. This bugfix adds a CSS variable to the graded group set container element that indicates the background color. The fade bars in the radio widget now reference this new background color variable, but falls back to the semantic color for backgrounds if a container color doesn't exist.

Issue: LEMS-3405

## Test plan:
1. Open Storybook (either for this PR, or locally)
1. Navigate to the new story: Widges => RadioNew => Docs => Scrolling in Graded Group Set
1. Note that when scrolling horizontally, the fade bars match the background color

Author: mark-fitzgerald

Reviewers: SonicScrewdriver, anakaren-rojas, ivyolamit

Required Reviewers:

Approved By: SonicScrewdriver, anakaren-rojas

Checks: ✅ 10 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2816