[image-regression-alignment] [Image] | (DX) | Fix alignment in regression stories
In order to fix the images' alignment in regression stories, I had to add
the width and height to the images in the testdata.

This is in line with real image data, as image widgets' `backgroundImage`s are
saved with the calculated widths and heights at publish.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3440

Test plan:
Storybook
`/?path=/docs/widgets-image-visual-regression-tests--docs`
- Confirm that all the images are center aligned

Check the UI Tests github check below and confirm that all the images have gone
from being left-aligned to being center-aligned.