Remove RenderProps & transform & staticTransform (#2864)
## Summary:

The reason I'm looking at RenderProps is because of the need for SSS to transition from an "answerless" to an "answerful" state - in order to provide hints, rationales, and answers to learners post-attempt. The problem is that Renderer stashes a copy of RenderProps (essentially the transformed output of passing widget options through either `transform` or `staticTransform`). Originally this problem manifested by resetting user input - when we went from answerless to answerful, Renderer considered it a new exercise and completely reset state. Once that was fixed the problem manifested in a different way - when going from answerless to answerful Renderer kept a copy of RenderProps which meant we never received the answerful data in the widgets.

> There are only two hard things in Computer Science: cache invalidation, naming things, and off-by-1 errors.

So what if there were no cache? That's the aim of this PR.

In previous PRs we:
- split UserInput into its own chunk of state
- removed all uses of `transform` and `staticTransform`
- internalized widget state within the widget itself (rather than in RenderProps)
- removed all "restore" serialized state uses / APIs (which depended on RenderProps)

This PR is mostly about cleaning up code. Since UserInput is stored separately, widgets store state internally, and there are no more transforms, RenderProps is just a copy of WidgetOptions now. With this PR we cut out the middle-man and pass WidgetOptions directly to the widgets removing the need for RenderProps.

Issues:
- [LEMS-3113](https://khanacademy.atlassian.net/browse/LEMS-3113) - Answerless Rendering trial run in Khan Classic
- [LEMS-3199](https://khanacademy.atlassian.net/browse/LEMS-3199) - WidgetOptions should be read only in the learner UX, deprecate transform/staticTransform
- [LEMS-3185](https://khanacademy.atlassian.net/browse/LEMS-3185) - Deprecate `getSerializedState` and `restoreSerializedState`

## Test plan:
- There should be no UX changes
- Unit tests should pass
- Doing a playtest before release

[LEMS-3199]: https://khanacademy.atlassian.net/browse/LEMS-3199?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[LEMS-3113]: https://khanacademy.atlassian.net/browse/LEMS-3113?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Author: handeyeco

Reviewers: handeyeco, benchristel, ivyolamit, mark-fitzgerald, SonicScrewdriver, anakaren-rojas, jeremywiebe

Required Reviewers:

Approved By: benchristel

Checks: ✅ 10 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2864