[benc/upgrade-chromaui-action] Upgrade chromaui/action to the latest version
This is an attempt to fix the problem of Chromatic UI tests hanging frequently
in CI.  I don't have any evidence that this will actually fix the problem, but
I'm not sure what else to try.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3533

Test plan:

CI, including the UI Tests, should pass on this PR without needing to be
kicked.