Support new division symbol (:) used by several locales (#2862)
## Summary:
This PR adds support for locales (particularly Ukraine) that use `:` instead of `/` for denoting division in math equations. We're only focusing on the Expression Widget as Numeric Input / Input Number do not support the division symbol. 

We're currently checking with the language advocates about how fractions are written in Ukrainian, but initial research is indicating that the bar `/` is still used in those cases. 

This PR does not update the actual icon on the keypad, as we're still waiting for design. We will update the keypad button (in the appropriate locales) in a future PR, after LEMS-3535 has been completed. 

## Video Example: 
https://github.com/user-attachments/assets/dd358910-ec21-4e76-8cc1-4405f59b5002

Issue: LEMS-3435

## Test plan:
- New tests pass 
- Manual testing using my [draft locale dropdown feature](https://github.com/Khan/perseus/pull/2861) 
- No regressions noticed in existing tests

Author: SonicScrewdriver

Reviewers: SonicScrewdriver, nishasy

Required Reviewers:

Approved By: nishasy

Checks: ✅ 10 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2862