[Radio] Bugfix - Ensure fade bars match the background color of container (#2928)
## Summary:
Widgets can be placed inside containers that have a background color that isn't communicated to the widget. In order for the radio widget to have scrolling fade bars match the background color of the container, it needs a way to get the background color. This PR adds a utility function that traverses the DOM ancestry of a given element until it finds a valid background color. It then uses it for the fade bar color. If no background color is found, then it uses the default background color for the theme.

Issue: LEMS-3545

## Test plan:
1. Open Storybook for this PR (found at the bottom of the checks - Storybook Publish)
2. Navigate to the Graded Group Set with Scroll regression story
   - Widgets => RadioNew => Visual Regression Tests => Static => Graded Group Set With Scroll
3. The fade area on the sides of the widget should match the background of the graded group container
   - Note: The first story on the page is not the story to inspect (it's merely a copy of the first story that Storybook automatically displays). The duplicate (first) story displayed will show white fade bars due to how Storybook implements that duplicate. The actual (second) story displayed will consistently show the fade bars with the same background as the container.

Author: mark-fitzgerald

Reviewers: mark-fitzgerald, ivyolamit, SonicScrewdriver, anakaren-rojas

Required Reviewers:

Approved By: SonicScrewdriver, ivyolamit

Checks: ✅ 10 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2928