[image-resize-zoom] [Image] | (UX) | Use large image size for zoom if available
We recently added the ability to resize an image in the editor - the new size
gets saved as part of the `backgroundImage.height` and `backgroundImage.width`
properties.

In the current implementation, this saved size is used for both rendering the
image at that size and for determining if it's a large enough resolution for
zooming. (The SVGImage only has the zoom feature if the image is larger than
its parent width.)

In this PR:
- Use the saved size for rendering the image within its item container
- Beteween the saved size and the original/natural image size, use the
  bigger one to determine zoom and "explore image" modal's image size.

This means that images can be rendered at a smaller size, even though they're
downloaded at full resolution, and this full resolution size can be used to
zoom into the image within the Image widget.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3389

Test plan:
Storybook
`/?path=/story/widgets-image-widget-demo--image-with-different-sizes`