[widget-delete-widget-list] Remove widget from item's `widgets` list when it's deleted from the editor
Right now, when a widget's "delete" button is pressed, it's deleted from the text editor,
but it's not deleted from the item's widgets list.

This doesn't seem right, and it may cause issues when trying to use the `widgets` list
to determine which widgets are in the content (like within the save warnings utility).

Updating the editor code so that deleting a widget also removes it from the widgets list.

Issue: none

Test plan:
Storybook
- Go to `http://localhost:6006/?path=/story/editors-editorpage--demo`
- Add a widget (e.g. radio)
- Click "Developer JSON Mode"
- Confirm the widget is in the JSON
- Click "Developer JSON Mode" again to toggle it off
- Delete the widget
- Toggle on JSON mode
- Confirm that the widget is no longer in the JSON