[Interactive Graph] Bugfix - Locked functions not rendering when domain is not set (#2995)
## Summary:
Infinity values are lost when they are serialized (i.e. via JSON). This PR adds a check for null values in the clampedDomain function to ensure proper functionality.

We discussed at length the prospects of having the parser conduct the validation logic. It was decided that since the likelihood of parsed data being serialized afterwards is high, then non-serializable values (like `Infinity`) should be validated within the widget. This decision is a two-way door, so if a different approach is preferred in the future, this should be easy to change out.

Issue: LEMS-3683

## Test plan:
1. Open Storybook
2. Navigate to Interactive Graph => Locked Functions
3. Note that all the stories are rendering their locked functions

Author: mark-fitzgerald

Reviewers: jeremywiebe, handeyeco

Required Reviewers:

Approved By: handeyeco, jeremywiebe

Checks: ✅ 17 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2995