[wb-comments-ticket] Refer to ticket LEMS-3686 in the comments for using CSS modules with WB
There are a number of TODO comments in the code saying to switch to CSS modules
once Wonder Blocks has support for it, but there was no ticket to point to.

I just created a ticket (LEMS-3686) and updated all the existing TODO comments
to refer to that ticket.

Issue: none

Test plan:
N/A