[fix-image-vertical-stretch] [Image] | (UX) | Fix image stretching vertically in markdown tables in Safari
The `height: 100%` style was causing the image to stretch verticaly as far as
it could, but only on Safari.

Removing this style seems to fix the issue without changing the image in any
of the other cases.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3691

Test plan:
Storybook
`/?path=/story/widgets-image-widget-demo--images-with-markdown-table`