[image-infinite-loop-bug] [Image] Possible fix for infinite loop rendering error
We've seen a recent uptick in the following error

Maximum update depth exceeded. This can happen when a component repeatedly calls setState inside componentWillUpdate or componentDidUpdate. React limits the number of nested updates to prevent infinite loops.

One possible cause for this could be passing in a whole object into the useEffect
dependency array rather than primitive values.

Updating that here in an attempt to reduce the rendering errors.

Issue: none

Test plan:
N/A (this isn't really easy to test)