Update the Interactive Graph Editor to use APIOptions to avoid a bug when saving (#3034)
## Summary:
When setting up the new `editingDisabled` feature, I didn't realize that the changeable `onChange` event was taking ALL of the sub-component props and passing them up. Unfortunately, this results in the `editingDisabled` property being serialized when it shouldn't be. 

In order to resolve this, I've removed the named property in favour of using the value from `apiOptions` directly — as we already have logic to strip out the `apiOptions` using a prop deny list. 

This makes for a more consistent experience, and is better set up for when we move `apiOptions` into a context. 

## Test plan:
- Current tests pass 
- Interactive graph can be updated and saved correctly in the main Khan Academy environment. (Verified)

Author: SonicScrewdriver

Reviewers: nishasy, Myranae

Required Reviewers:

Approved By: nishasy

Checks: ⏭️  1 check has been skipped, ✅ 10 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/3034