[Image] | (DX) | Remove ZoomService in favor of WB Modal (#3009)
## Summary:
The old zoom service is a huge pain to work with. It's super old and very
non-modern-reacty.

I'm replacing it here with Wonder Blocks modal so that the zoom feature can be
easier to work with and maintain.

This is a prerequisite for making the graphies focusable and zoomable since the
old zoom service was trying to do too much with the image ref, making it impossible
to zoom in on graphies with their labels.

NOTE: The nice little transition is unfortunately gone with this approach. But I got
confirmation from our designer that we can go with this for now, and we can add
in transitions once WB Modal supports it next year.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3689

## Test plan:
Storybook
`/?path=/story/widgets-image-widget-demo--image-with-different-sizes`

### Before

https://github.com/user-attachments/assets/509c0a4b-9ee9-48de-9966-8fb6db6189e5

### After

https://github.com/user-attachments/assets/e4c13228-4028-4bce-9371-b5f30dbcba6e

Author: nishasy

Reviewers: nishasy, ivyolamit, catandthemachines, jeremywiebe, SonicScrewdriver, mark-fitzgerald, anakaren-rojas

Required Reviewers:

Approved By: ivyolamit, catandthemachines

Checks: ⏭️  1 check has been skipped, ✅ 10 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/3009