Preview like production (#3018)
## Summary:

This PR introduces a new preview setup for Storybook. It mimics exactly what production docs and as a result, should make for more accurate stories for Editor-side components. 

It uses a trick that I noticed with Storybook where you can get a "deep-link" directly to a given story. Using this, I've built a "preview host" Story that acts as the page which renders the preview content. Building on this, we then point the `<iframe>` preview url to this Story's direct link and voila, we have a working production-like preview. Further, because we're using a Story to mimic the content preview page, we can edit and work with it just like any other Typescript code in our system (ie. we have easy page-reload refreshes and full type safety). 

This PR now paves the way to re-work our iframe-based preview system to be much more reliable and approachable as we'll be able to refactor both sides of the preview _and_ build the components for both sides right here in Storybook instead of having the functionality split between this repo and our frontend repo. 



Issue: LEMS-3741

## Test plan:

Run `pnpm start`
Navigate to the `EditorPage` stories and edit some content. You should see the preview updating as you edit and if you inspect the DOM, you should see the <iframe> pointing to the location Storybook instance.

Author: jeremywiebe

Reviewers: jeremywiebe, benchristel, nishasy, catandthemachines, ivyolamit

Required Reviewers:

Approved By: benchristel

Checks: ⏭️  1 check has been skipped, ✅ 9 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/3018