[khanmigo-server-side] Move the Khanmigo data functions to Perseus Core for Server Side Scoring initiative
This PR moves getAnswersFromWidget, injectWidgets, and a handful of helper functions to Perseus Core so that they are available for the Perseus service. We will be creating a new API endpoint to call these functions, so that we can support Khanmigo's move to keeping answerful data server-side.

It also created the new getPerseusAIData function, which will be used for the new API Endpoint.

Issue: LEMS-3766

Test plan:
Tests pass